fix: make this shit work

+++ linuxdoom-1.10/r_draw.c
@@ -460,8 +460,8 @@ void R_InitTranslationTables (void)
 {
     int		i;
 	
-    translationtables = Z_Malloc (256*3+255, PU_STATIC, 0);
-    translationtables = (byte *)(( (int)translationtables + 255 )& ~255);
+	translationtables = Z_Malloc (256*3+255, PU_STATIC, 0);
+	translationtables = (byte *)(( (intptr_t)translationtables + 255 )& ~255);
     
     // translate just the 16 green colors
     for (i=0 ; i<256 ; i++)
